Restructured the Andorid project into Packages
Make a package per big subgroup!
