+++ client/android/Crowd-ML/app/src/main/AndroidManifest.xml
@@ -7,7 +7,7 @@
     <uses-permission android:name="android.permission.WAKE_LOCK" />
 
     <application
-        android:name=".CrowdMLApplication"
+        android:name=".firebase.CrowdMLApplication"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
@@ -50,8 +50,8 @@
             android:authorities="@string/multiprocess_preferences_authority"
             android:exported="false" />
         <provider
-            android:name=".FirebaseProvider"
-            android:authorities="osu.crowd_ml.FirebaseProvider"
+            android:name=".firebase.FirebaseProvider"
+            android:authorities="osu.crowd_ml.firebase.FirebaseProvider"
             android:process=":datasend"
             android:exported="false"
             android:syncable="false" />
