+++ client/android/Crowd-ML/app/src/main/java/osu/crowd_ml/OfflineDataSend.java
@@ -1,7 +1,5 @@
 package osu.crowd_ml;
 
-import android.content.Context;
-import android.content.Intent;
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
 import android.view.View;
@@ -17,6 +15,15 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
+
+import osu.crowd_ml.loss_functions.LossFunction;
+import osu.crowd_ml.loss_functions.Softmax;
+import osu.crowd_ml.noise_distributions.Distribution;
+import osu.crowd_ml.noise_distributions.NoNoise;
+import osu.crowd_ml.tests.ModelTest;
+import osu.crowd_ml.tests.BinaryTest;
+import osu.crowd_ml.tests.MultiTest;
+import osu.crowd_ml.tests.NNTest;
 
         /*
         Copyright 2016 Crowd-ML team
@@ -69,7 +76,7 @@ public class OfflineDataSend extends AppCompatActivity {
     List<Integer> testLabels = new ArrayList<Integer>(testN);
 
     private int length;
-    private AccTest test;
+    private ModelTest test;
