+++ client/android/Crowd-ML/app/src/main/java/osu/crowd_ml/loss_functions/MultiSVM.java
@@ -1,7 +1,9 @@
-package osu.crowd_ml;
+package osu.crowd_ml.loss_functions;
 
 import java.util.ArrayList;
 import java.util.List;
+
+import osu.crowd_ml.loss_functions.LossFunction;
 
 /*
     Copyright 2016 Crowd-ML team
@@ -20,7 +22,7 @@ import java.util.List;
     limitations under the License
 */
 
-public class MultiSVM implements LossFunction{
+public class MultiSVM implements LossFunction {
 
     public String lossType() {
         return "multi";
