+++ client/android/Crowd-ML/app/src/main/java/osu/crowd_ml/loss_functions/SoftmaxNN.java
@@ -1,7 +1,9 @@
-package osu.crowd_ml;
+package osu.crowd_ml.loss_functions;
 
 import java.util.ArrayList;
 import java.util.List;
+
+import osu.crowd_ml.loss_functions.LossFunction;
 
 /*
 Copyright 2016 Crowd-ML team
